# openapi.yaml
openapi: 3.0.1
info:
  title: Metadata Standards Catalog
  description: "The RDA [Metadata Standards Catalog](https://rdamsc.bath.ac.uk/) is a collaborative, open directory of metadata standards applicable to research data. It is offered to the international academic community to help address infrastructure challenges."
  termsOfService: https://rdamsc.bath.ac.uk/terms-of-use
  version: 2.0.0
...
servers:
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/alex-ball/rda-metadata-standards-catalog/2.0.0
- description: Live server
  url: https://rdamsc.bath.ac.uk
...
paths:
...
  /api2/rel:
    get:
      tags:
      - rel
      summary: "List all entity relation records"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsePageRelation'
    post:
      tags:
      - rel
      summary: "Submit a new entity relation record"
      requestBody:
        description: "Complete record in JSON."
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Relation'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSuccessRelation'
        400:
          description: Failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponseFailure'
        401:
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - api_user: []
...
  /api2/user/reset-password:
    post:
      tags:
      - user
      summary: "Change API user password"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                new_password:
                  type: string
                  description: "New password to use for subsequent calls."
              required:
              - new_password
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsePassword'
        400:
          description: "Bad Request: no suitable password provided"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsePassword'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        500:
          description: "Internal Server Error: password could not be reset"
          content: {}
      security:
      - api_user: []
  /api2/user/token:
    get:
      tags:
      - user
      summary: "Request an authorization token"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiVersion:
                    type: string
                  token:
                    type: string
                    description: "Authorization token valid for 600 seconds."
        401:
          $ref: '#/components/responses/UnauthorizedError'
      security:
      - api_user: []
components:
  schemas:
...
    MSCID:
      type: string
      readOnly: true
      description: "Non-semantic internal identifier for the record."
      pattern: '^msc:(m|g|t|c|e|datatype|location|type|id_scheme)\d+$'
...
    Relation:
      type: object
      properties:
        "@id":
          $ref: '#/components/schemas/MSCID'
        parent schemes:
          type: array
          description: "Schemes of which the current scheme is a profile. In practice, this means either that the current scheme explicitly “borrows” elements from the parent scheme, or that all instances of the current scheme would be valid instances of the parent scheme."
          items:
            type: string
            description: "Non-semantic internal identifier for the related record."
            pattern: '^msc:m\d+$'
        supported schemes:
          type: array
          description: "Schemes to which the tool’s input metadata should conform, or output metadata does conform."
          items:
            type: string
            description: "Non-semantic internal identifier for the related record."
            pattern: '^msc:m\d+$'
        input schemes:
          type: array
          description: "Schemes to which a metadata record input to the current mapping must conform."
          items:
            type: string
            description: "Non-semantic internal identifier for the related record."
            pattern: '^msc:m\d+$'
        output schemes:
          type: array
          description: "Schemes to which metadata records output from the current mapping conform."
          items:
            type: string
            description: "Non-semantic internal identifier for the related record."
            pattern: '^msc:m\d+$'
        endorsed schemes:
          type: array
          description: "Schemes endorsed by the current endorsement."
          items:
            type: string
            description: "Non-semantic internal identifier for the related record."
            pattern: '^msc:m\d+$'
        maintainers:
          type: array
          description: "Organizations that maintain the current scheme, tool or mapping."
          items:
            type: string
            description: "Non-semantic internal identifier for the related record."
            pattern: '^msc:g\d+$'
        funders:
          type: array
          description: "Organizations that fund the current scheme, tool or mapping."
          items:
            type: string
            description: "Non-semantic internal identifier for the related record."
            pattern: '^msc:g\d+$'
        users:
          type: array
          description: "Organizations that are known to use the current scheme."
          items:
            type: string
            description: "Non-semantic internal identifier for the related record."
            pattern: '^msc:g\d+$'
        originators:
          type: array
          description: "Organizations that have made the current endorsement."
          items:
            type: string
            description: "Non-semantic internal identifier for the related record."
            pattern: '^msc:g\d+$'
...
    Response:
      type: object
      properties:
        apiVersion:
          type: string
...
    Page:
      type: object
      properties:
        itemsPerPage:
          type: integer
          description: "Number of items per page of results."
        currentItemCount:
          type: integer
          description: "Number of items in the current page of results."
        startIndex:
          type: integer
          description: "Index of the first item in the current page of results. Items are counted from 1."
        totalItems:
          type: integer
          description: "Total number of items in the result set (equal to the index of the last item)."
        pageIndex:
          type: integer
          description: "Index of the current page of results. Pages are counted from 1."
        totalPages:
          type: integer
          description: "Total number of pages in the result"
        nextLink:
          type: string
          description: "URL from which to obtain the next page of results."
        previousLink:
          type: string
          description: "URL from which to obtain the previous page of results."
...
    ResponsePageRelation:
      allOf:
      - $ref: '#/components/schemas/Response'
      - properties:
          data:
            allOf:
              - $ref: '#/components/schemas/Page'
              - properties:
                  items:
                    type: array
                    description: "Items in the current page of results."
                    items:
                      $ref: '#/components/schemas/Relation'
...
    ResponseSuccessRelation:
      allOf:
      - $ref: '#/components/schemas/Response'
      - properties:
          data:
            $ref: '#/components/schemas/Relation'
...
    SubmissionResponseFailure:
      type: object
      properties:
        apiVersion:
          type: string
        error:
          type: object
          properties:
            message:
              type: string
              description: "Summary of the error(s)."
            errors:
              type: array
              description: "Errors returned by form validation."
              items:
                type: object
                properties:
                  message:
                    type: string
                    description: "Nature of validation error."
                  location:
                    type: string
                    description: "Object key where the error occurred."
    ResponsePassword:
      type: object
      properties:
        apiVersion:
          type: string
        username:
          type: string
          description: "Confirmation of the username for which the password was reset."
        password_reset:
          type: boolean
          description: "True if the reset was successful, false otherwise."
  responses:
    UnauthorizedError:
      description: Unauthorized access
      headers:
        WWW_Authenticate:
          schema:
            type: string
  securitySchemes:
    api_user:
      type: http
      description: "If using an API token, use this as the username and any arbitrary string as the password."
      scheme: basic